review: simplified paperHero calls, type PaperHeroItems added, gets used and seperate style and copy files added
